+++ src/main/java/com/choogoomoneyna/choogoomoneyna_be/auth/jwt/util/JwtProperties.java
@@ -1,4 +1,4 @@
-package com.choogoomoneyna.choogoomoneyna_be.auth.jwt;
+package com.choogoomoneyna.choogoomoneyna_be.auth.jwt.util;
 
 import org.springframework.beans.factory.annotation.Value;
 import lombok.Getter;
